+++ src/rinetd.h
@@ -18,8 +18,9 @@ static int const RINETD_LISTEN_BACKLOG = 128;
 static int const RINETD_DEFAULT_UDP_TIMEOUT = 10;
 static int const RINETD_MAX_UDP_CONNECTIONS = 1000;
 
-#define RINETD_CONFIG_FILE "/etc/rinetd.conf"
-#define RINETD_PID_FILE "/var/run/rinetd.pid"
+#define RINETD_CONFIG_FILE "/etc/rinetd-uv.conf"
+#define RINETD_PID_FILE "/var/run/rinetd-uv.pid"
+#define MAX_INCLUDE_DEPTH 10
 
 /* Global configuration */